#821 [REFACTOR] 폴더 구조 정리(숙명인 인증) - page/account
