+++ src/feature/account/component/CompleteStep/CompleteStep.jsx
@@ -1,19 +1,19 @@
 import { Link } from 'react-router-dom';
 
-import { Button } from '@/pages/SnoroseVerifyPage';
+import { PrimaryButton } from '@/shared/component';
 
 import Complete from '@/assets/images/completeVerification.svg';
 
-import styles from './CompletePage.module.css';
+import styles from './CompleteStep.module.css';
 
-export default function CompletePage() {
+export default function CompleteStep() {
   return (
     <section className={styles.content}>
       <div className={styles.illust}>
         <img src={Complete} alt='인증 신청 완료' />
       </div>
       <Link to='/'>
-        <Button>메인 페이지로</Button>
+        <PrimaryButton>메인 페이지로</PrimaryButton>
       </Link>
     </section>
   );
