+++ src/feature/account/component/index.js
@@ -1,1 +1,7 @@
+// 숙명인 인증
+export { default as TermsStep } from './TermsStep/TermsStep';
+export { default as VerifyStep } from './VerifyStep/VerifyStep';
+export { default as CompleteStep } from './CompleteStep/CompleteStep';
+
+export { default as AuthInput } from './AuthInput/AuthInput';
 export { default as StageDots } from './StageDots/StageDots';
