+++ src/pages/NotFoundPage/NotFoundPage.jsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
 
-import { Button } from '@/pages/SnoroseVerifyPage';
+import { PrimaryButton } from '@/shared/component';
 
 import styles from './NotFoundPage.module.css';
 
@@ -10,7 +10,7 @@ export default function NotFoundPage() {
       <h1 className={styles.error}>404</h1>
       <div className={styles.text}>페이지를 찾을 수 없습니다.</div>
       <Link className={styles.button} to='/'>
-        <Button>메인으로</Button>
+        <PrimaryButton>메인으로</PrimaryButton>
       </Link>
     </main>
   );
