--- src/pages/SnoroseVerifyPage/index.js
@@ -1,6 +1,0 @@
-export { default as SnoroseVerifyPage } from './SnoroseVerifyPage';
-export { default as TermsPage } from './TermsPage/TermsPage';
-export { default as VerifyPage } from './VerifyPage/VerifyPage';
-export { default as CompletePage } from './CompletePage/CompletePage';
-export { default as Input } from './Input/Input';
-export { default as Button } from './Button/Button';
